requirements.txt
pytest==8.2.1
selenium==4.21.0
python-dotenv==1.0.1
allure-pytest==2.13.5
pytest-xdist==3.6.1